+++ packages/block-library/src/file/index.js
@@ -6,7 +6,7 @@ import { includes } from 'lodash';
 /**
  * WordPress dependencies
  */
-import { __ } from '@wordpress/i18n';
+import { __, _x } from '@wordpress/i18n';
 import { createBlobURL } from '@wordpress/blob';
 import { createBlock } from '@wordpress/blocks';
 import { select } from '@wordpress/data';
@@ -39,6 +39,7 @@ export const settings = {
 			type: 'string',
 		},
 		fileName: {
+			type: 'string',
 			source: 'html',
 			selector: 'a:not([download])',
 		},
@@ -61,9 +62,10 @@ export const settings = {
 			default: true,
 		},
 		downloadButtonText: {
+			type: 'string',
 			source: 'html',
 			selector: 'a[download]',
-			default: __( 'Download' ),
+			default: _x( 'Download', 'button label' ),
 		},
 	},