+++ packages/block-library/src/subhead/index.js
@@ -31,6 +31,7 @@ export const settings = {
 
 	attributes: {
 		content: {
+			type: 'string',
 			source: 'html',
 			selector: 'p',
 		},