+++ packages/blocks/src/api/test/registration.js
@@ -16,8 +16,6 @@ import { addFilter, removeFilter } from '@wordpress/hooks';
 import {
 	registerBlockType,
 	unregisterBlockType,
-	setUnknownTypeHandlerName,
-	getUnknownTypeHandlerName,
 	setFreeformContentHandlerName,
 	getFreeformContentHandlerName,
 	setUnregisteredTypeHandlerName,
@@ -31,6 +29,7 @@ import {
 	isReusableBlock,
 	unstable__bootstrapServerSideBlockDefinitions, // eslint-disable-line camelcase
 	registerBlockStyle,
+	unregisterBlockStyle,
 } from '../registration';
 
 describe( 'blocks', () => {
@@ -371,29 +370,6 @@ describe( 'blocks', () => {
 		} );
 	} );
 
-	describe( 'setUnknownTypeHandlerName()', () => {
-		it( 'assigns unknown type handler', () => {
-			try {
-				setUnknownTypeHandlerName( 'core/test-block' );
-
-				expect( getUnknownTypeHandlerName() ).toBe( 'core/test-block' );
-				expect( console ).toHaveWarned();
-			} finally {
-				// Restore undefined handler here rather than in `afterEach` because:
-				// - This call generates a deprecation warning.
-				// - Deprecation warnings become test errors unless we assert `toHaveWarned`.
-				// - This is too broad of an assertion to apply for all tests in the suite.
-				setUnknownTypeHandlerName( undefined );
-			}
-		} );
-	} );
-
-	describe( 'getUnknownTypeHandlerName()', () => {
-		it( 'defaults to undefined', () => {
-			expect( getUnknownTypeHandlerName() ).toBeNull();
-		} );
-	} );
-
 	describe( 'setFreeformContentHandlerName()', () => {
 		it( 'assigns unknown type handler', () => {
 			setFreeformContentHandlerName( 'core/test-block' );
@@ -636,6 +612,59 @@ describe( 'blocks', () => {
 				{ name: 'normal', label: 'Normal style' },
 				{ name: 'small', label: 'Small style' },
 				{ name: 'big', label: 'Big style' },
+			] );
+		} );
+	} );
+
+	describe( 'unregisterBlockStyle', () => {
+		afterEach( () => {
+			removeFilter( 'blocks.registerBlockType', 'my-plugin/block-with-styles/big/unregister' );
+			removeFilter( 'blocks.registerBlockType', 'my-plugin/block-with-styles/small/unregister' );
+			removeFilter( 'blocks.registerBlockType', 'my-plugin/block-with-styles/big' );
+			removeFilter( 'blocks.registerBlockType', 'my-plugin/block-with-styles/small' );
+		} );
+
+		it( 'should remove styles', () => {
+			unregisterBlockStyle( 'my-plugin/block-with-styles', 'big' );
+			const settings = registerBlockType( 'my-plugin/block-with-styles', {
+				...defaultBlockSettings,
+				styles: [ { name: 'big', label: 'Big style' } ],
+			} );
+
+			expect( settings.styles ).toEqual( [] );
+		} );
+
+		it( 'should keep other styles', () => {
+			unregisterBlockStyle( 'my-plugin/block-with-styles', 'small' );
+			const settings = registerBlockType( 'my-plugin/block-with-styles', {
+				...defaultBlockSettings,
+				styles: [
+					{ name: 'normal', label: 'Normal style' },
+					{ name: 'small', label: 'Small style' },
+					{ name: 'big', label: 'Big style' },
+				],
+			} );
+
+			expect( settings.styles ).toEqual( [
+				{ name: 'normal', label: 'Normal style' },
+				{ name: 'big', label: 'Big style' },
+			] );
+		} );
+
+		it( 'should remove a prior registerBlockStyle', () => {
+			registerBlockStyle( 'my-plugin/block-with-styles', { name: 'big', label: 'Big style' } );
+			registerBlockStyle( 'my-plugin/block-with-styles', { name: 'small', label: 'Small style' } );
+			unregisterBlockStyle( 'my-plugin/block-with-styles', 'big' );
+			const settings = registerBlockType( 'my-plugin/block-with-styles', {
+				...defaultBlockSettings,
+				styles: [
+					{ name: 'normal', label: 'Normal style' },
+				],
+			} );
+
+			expect( settings.styles ).toEqual( [
+				{ name: 'normal', label: 'Normal style' },
+				{ name: 'small', label: 'Small style' },
 			] );
 		} );
 	} );