+++ packages/editor/src/components/block-drop-zone/index.js
@@ -1,7 +1,6 @@
 /**
  * External Dependencies
  */
-import { castArray } from 'lodash';
 import classnames from 'classnames';
 
 /**
@@ -10,7 +9,6 @@ import classnames from 'classnames';
 import { DropZone } from '@wordpress/components';
 import {
 	rawHandler,
-	cloneBlock,
 	getBlockTransforms,
 	findTransform,
 } from '@wordpress/blocks';
@@ -132,17 +130,7 @@ export default compose(
 
 		return {
 			insertBlocks( blocks, index ) {
-				const { rootClientId, layout } = ownProps;
-
-				if ( layout ) {
-					// A block's transform function may return a single
-					// transformed block or an array of blocks, so ensure
-					// to first coerce to an array before mapping to inject
-					// the layout attribute.
-					blocks = castArray( blocks ).map( ( block ) => (
-						cloneBlock( block, { layout } )
-					) );
-				}
+				const { rootClientId } = ownProps;
 
 				insertBlocks( blocks, index, rootClientId );
 			},
@@ -150,8 +138,8 @@ export default compose(
 				updateBlockAttributes( ...args );
 			},
 			moveBlockToPosition( srcClientId, srcRootClientId, dstIndex ) {
-				const { rootClientId: dstRootClientId, layout } = ownProps;
-				moveBlockToPosition( srcClientId, srcRootClientId, dstRootClientId, layout, dstIndex );
+				const { rootClientId: dstRootClientId } = ownProps;
+				moveBlockToPosition( srcClientId, srcRootClientId, dstRootClientId, dstIndex );
 			},
 		};
 	} ),