+++ packages/blocks/CHANGELOG.md
@@ -1,3 +1,25 @@
+## 5.1.0 (Unreleased)
+
+### New feature
+
+- `isValidBlockContent` function has been added ([#10891](https://github.com/WordPress/gutenberg/pull/10891)).
+
+### Deprecation
+
+- `isValidBlock` function has been deprecated ([#10891](https://github.com/WordPress/gutenberg/pull/10891)). Use `isValidBlockContent` instead.
+
+## 5.0.0 (2018-10-29)
+
+### Breaking Changes
+
+- Attribute type coercion has been removed. Omit the source to preserve type via serialized comment demarcation.
+- `setUnknownTypeHandlerName` has been removed. Please use `setFreeformContentHandlerName` and `setUnregisteredTypeHandlerName` instead.
+- `getUnknownTypeHandlerName` has been removed. Please use `getFreeformContentHandlerName` and `getUnregisteredTypeHandlerName` instead.
+
+### New Feature
+
+- Added a `unregisterBlockStyle()` function to remove a block style variation.
+
 ## 4.0.4 (2018-10-19)
 
 ## 4.0.3 (2018-10-18)